Fix dispatch refactor issues
