+++ src/compiler/scope_analysis.c
@@ -1,3 +1,4 @@
+#define _POSIX_C_SOURCE 200809L
 #include "../../include/vm.h"
 #include "../../include/compiler.h"
 #include <stdlib.h>
